Refactor argparsing for perfomance
It can be beneficial to avoid top-level imports until after the CLI
arguments have been parsed. If `--help` was given by the user, the
program will exit anyway, so expensive imports should be deferred.

In this case, it was necessary to split markitdown into two packages:
library and application. The reason for this is that markitdown exports
some expressions in `__init__.py` before anything else happens.

Before:

    $ time .venv/bin/markitdown -h > /dev/null

    ________________________________________________________
    Executed in  240.48 millis    fish           external
       usr time    1.27 secs    726.00 micros    1.27 secs
       sys time    0.04 secs    367.00 micros    0.04 secs

After:

    $ time .venv/bin/markitdown -h > /dev/null

    ________________________________________________________
    Executed in   30.75 millis    fish           external
       usr time   14.88 millis    0.00 millis   14.88 millis
       sys time   16.01 millis    1.13 millis   14.88 millis